requirements.txt
pandas==2.1.3
torch==2.2.0.dev20231122+cu121
sklearn==0.0.post10
numpy==1.24.1
matplotlib==3.8.0
tqdm==4.66.1
pillow==9.3.0
torchvision==0.17.0.dev20231122+cu121
opencv-python==4.8.1.78